DB procedures for creating mysql database added, separated user db crud operations and product db crud operations
