--- app/CustomsControlSM/Controllers/FrontControllerInterface.php
@@ -1,7 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: stojan.mitric
- * Date: 4/24/2017
- * Time: 2:46 PM
- */
