+++ app/CustomsControlSM/Controllers/ProductEvidenceController.php
@@ -1,19 +1,17 @@
 <?php
+
 namespace CustomsControlSM\Controllers;
 
 use CustomsControlSM\Model\ProductEvidence;
 
-class ProductEvidenceController implements FrontControllerInterface
+class ProductEvidenceController
 {
 	private $productEvidence;
 
-	const DEFAULT_CONTROLLER = "IndexController";
-	const DEFAULT_ACTION     = "index";
+	public function __construct(array $options = array()) {
 
-	protected $controller    = self::DEFAULT_CONTROLLER;
-	protected $action        = self::DEFAULT_ACTION;
-	protected $params        = array();
-	protected $basePath      = "mybasepath/";
+		$this->productEvidence = new ProductEvidence();
+	}
 
 	public function save()
 	{
@@ -44,77 +42,9 @@ class ProductEvidenceController implements FrontControllerInterface
 
 	public function update()
 	{
-
 		//update file
 	}
 
 
-
-	public function __construct(array $options = array()) {
-
-		$this->productEvidence = new ProductEvidence();
-
-		if (empty($options)) {
-			$this->parseUri();
-		}
-		else {
-			if (isset($options["controller"])) {
-				$this->setController($options["controller"]);
-			}
-			if (isset($options["action"])) {
-				$this->setAction($options["action"]);
-			}
-			if (isset($options["params"])) {
-				$this->setParams($options["params"]);
-			}
-		}
-	}
-
-	protected function parseUri() {
-		$path = trim(parse_url($_SERVER["REQUEST_URI"], PHP_URL_PATH), "/");
-		$path = preg_replace('/[^a-zA-Z0-9]//', "", $path);
-		if (strpos($path, $this->basePath) === 0) {
-			$path = substr($path, strlen($this->basePath));
-		}
-		@list($controller, $action, $params) = explode("/", $path, 3);
-		if (isset($controller)) {
-			$this->setController($controller);
-		}
-		if (isset($action)) {
-			$this->setAction($action);
-		}
-		if (isset($params)) {
-			$this->setParams(explode("/", $params));
-		}
-	}
-
-	public function setController($controller) {
-		$controller = ucfirst(strtolower($controller)) . "Controller";
-		if (!class_exists($controller)) {
-			throw new InvalidArgumentException(
-				"The action controller '$controller' has not been defined.");
-		}
-		$this->controller = $controller;
-		return $this;
-	}
-
-	public function setAction($action) {
-		$reflector = new ReflectionClass($this->controller);
-		if (!$reflector->hasMethod($action)) {
-			throw new InvalidArgumentException(
-				"The controller action '$action' has been not defined.");
-		}
-		$this->action = $action;
-		return $this;
-	}
-
-	public function setParams(array $params) {
-		$this->params = $params;
-		return $this;
-	}
-
-	public function run() {
-		call_user_func_array(array(new $this->controller, $this->action), $this->params);
-	}
 }
 ?>
