--- app/CustomsControlSM/db/IDatabase.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace CustomsControlSM\db;
-
-interface IDatabase {
-
-	public function create($user, $fileName);
-
-	public function update($user,$file);
-
-	public function delete($user, $file);
-
-	public function listAll();
-
-}
-
-?>
