+++ app/CustomsControlSM/db/IUserCRUD.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace CustomsControlSM\db;
+
+interface IUserCRUD {
+
+	public function create();
+
+	public function update();
+
+	public function delete();
+
+	public function listAll();
+
+}
+
+?>
