+++ app/CustomsControlSM/db/UserMongoCRUD.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace CustomsControlSM\db;
+
+class UserMongoCRUD implements IUserCRUD {
+
+	public function create();
+
+	public function update();
+
+	public function delete();
+
+	public function listAll();
+}
+
+?>
