+++ config.php
@@ -1,4 +1,5 @@
 <?php
 
+define('ACTIVE_DB', 'mysql');
 
 ?>
