+++ db_procedures.sql
@@ -1,0 +1,9 @@
+CREATE DATABASE `orders`;
+
+CREATE TABLE `orders`.`users` (
+  `id` INT NOT NULL AUTO_INCREMENT,
+  `name` VARCHAR(45) NULL,
+  `email` VARCHAR(45) NULL,
+  `username` VARCHAR(45) NULL,
+  `password` VARCHAR(45) NULL);
+);
