+++ public/index.php
@@ -1,15 +1,4 @@
 <?php
-
-require_once "../config.php";
-
-require '../vendor/autoload.php';
-
-$loader = new Twig_Loader_Filesystem('public');
-$twig = new Twig_Environment($loader);
-
-echo $twig->render('index.html', array(
-    'name' => 'stole',
-    'age' => 25
-));
+require_once "../bootstrap/bootstrap.php";
 
 ?>
