--- public/templates/partials/messages.php
@@ -1,3 +1,0 @@
-{% if flash.global %}
-	<div class ="global">{{ flash.global }}</div>
-{% endif %}
