--- public/templates/partials/navigation.php
@@ -1,9 +1,0 @@
-
-<ul>
-        <li> <a href="{{ urlFor('home') }}"> Home </a> </li>
-        <li> <a href="{{ urlFor('register') }}"> Register </a> </li>
-        <li> <a href="{{ urlFor('login') }}"> Login </a> </li>
-        <li> <a href="{{ urlFor('user.all') }}"> All users </a> </li>
-        <li> <a href="{{ urlFor('purchasedProducts') }}"> Purchased products </a> </li>
-
-</ul>
